fix type and replace intermediate_steps
